+++ src/test/java/janggi/service/GameServiceTest.java
@@ -3,7 +3,7 @@ package janggi.service;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import javax.sql.DataSource;
-import janggi.TestDataInitializer;
+import janggi.config.TestDataInitializer;
 import janggi.config.TestDataSourceConfig;
 import janggi.domain.Position;
 import janggi.domain.board.Board;
